fix caching, alerts menu...
fix for api calls that return very large amounts of data. plex's JSON to object has a hard limit on the amount of data it will work with. Sonarr makes you get all episodes for a show at a time, not just a season, so you can get >250kb worth of data in a call.